requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

numpy
pandas
scikit-learn
matplotlib
tqdm
pillow
opencv-python
seaborn